Added windows support

+++ src/poolallocationtracker.cpp
@@ -9,7 +9,7 @@ namespace MCR
 	
 	PoolAllocationTracker::FindAvailableResult PoolAllocationTracker::FindAvailable(uint64_t elementCount)
 	{
-		ssize_t blockIndex = -1;
+		long blockIndex = -1;
 		
 		for (size_t i = 0; i < m_availableBlocks.size(); i++)
 		{
@@ -47,8 +47,8 @@ namespace MCR
 	
 	void PoolAllocationTracker::Free(uint64_t firstElement, uint64_t elementCount)
 	{
-		ssize_t prevBlockIndex = -1;
-		ssize_t nextBlockIndex = -1;
+		long prevBlockIndex = -1;
+		long nextBlockIndex = -1;
 		
 		const uint64_t nextBlockFirstElement = firstElement + elementCount;
